[BUGFIX] Fix active page calculation in indexed_search
Because of introducing a strict comparion in #60346, the active page is
not highlighted anymore as a float and an int value are compared.

Resolves: #62532
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/33606
Tested-by: Georg Ringer <georg.ringer@gmail.com>

+++ typo3/sysext/indexed_search/Classes/Controller/SearchFormController.php
@@ -1609,7 +1609,7 @@ class SearchFormController extends \TYPO3\CMS\Frontend\Plugin\AbstractPlugin {
 				// all pages after the 1st one
 				$links[] = '<li>' . $this->makePointerSelector_link($this->pi_getLL('pi_list_browseresults_prev', '< Previous', TRUE), $pointer - 1, $freeIndexUid) . '</li>';
 			}
-			$minPage = $pointer - floor($maxPages / 2);
+			$minPage = $pointer - (int)floor($maxPages / 2);
 			$maxPage = $minPage + $maxPages - 1;
 			// Check if the indexes are within the page limits
 			if ($minPage < 0) {
